Silenced two clang warnings
Former-commit-id: 1bc400d850d1272178b0a25640059e1dfb71b71a

+++ src/game/Player.cpp
@@ -21926,7 +21926,7 @@ bool Player::IsInDuelArea() const
 
 void Player::SetDifficulty(Difficulty dungeon_difficulty, bool sendToPlayer, bool asGroup)
 {
-    ASSERT(dungeon_difficulty < MAX_DIFFICULTY);
+    ASSERT(uint32(dungeon_difficulty) < MAX_DIFFICULTY);
     m_dungeonDifficulty = dungeon_difficulty;
     if (sendToPlayer)
         SendDungeonDifficulty(asGroup);
